+++ crates/holochain/src/conductor/conductor.rs
@@ -129,6 +129,8 @@ mod graft_records_onto_source_chain;
 
 mod app_auth_token_store;
 
+mod hc_p2p_handler_impl;
+
 /// Verify signature of a signed zome call.
 ///
 /// [Signature verification](holochain_conductor_api::AppRequest::CallZome)
@@ -856,11 +858,9 @@ mod network_impls {
     use holochain_conductor_api::{
         CellInfo, DnaStorageInfo, NetworkInfo, StorageBlob, StorageInfo,
     };
-    use holochain_p2p::HolochainP2pResult;
     use holochain_sqlite::stats::{get_size_on_disk, get_used_size};
     use holochain_zome_types::block::Block;
     use holochain_zome_types::block::BlockTargetId;
-    use kitsune2_api::BoxFut;
     use zome_call_signature_verification::is_valid_signature;
 
     use crate::conductor::api::error::{
@@ -1448,106 +1448,6 @@ mod network_impls {
                 Ok(Err(error)) => Err(ConductorApiError::Other(Box::new(error))),
                 Err(error) => Err(error),
             }
-        }
-    }
-
-    impl holochain_p2p::event::HcP2pHandler for Conductor {
-        fn handle_call_remote(
-            &self,
-            dna_hash: DnaHash,
-            to_agent: AgentPubKey,
-            zome_call_params_serialized: ExternIO,
-            signature: Signature,
-        ) -> BoxFut<'_, HolochainP2pResult<SerializedBytes>> {
-            Box::pin(async { unimplemented!() })
-        }
-
-        fn handle_publish(
-            &self,
-            dna_hash: DnaHash,
-            request_validation_receipt: bool,
-            countersigning_session: bool,
-            ops: Vec<holochain_types::dht_op::DhtOp>,
-        ) -> BoxFut<'_, HolochainP2pResult<()>> {
-            Box::pin(async { unimplemented!() })
-        }
-
-        fn handle_get(
-            &self,
-            dna_hash: DnaHash,
-            to_agent: AgentPubKey,
-            dht_hash: holo_hash::AnyDhtHash,
-            options: holochain_p2p::event::GetOptions,
-        ) -> BoxFut<'_, HolochainP2pResult<WireOps>> {
-            Box::pin(async { unimplemented!() })
-        }
-
-        fn handle_get_meta(
-            &self,
-            dna_hash: DnaHash,
-            to_agent: AgentPubKey,
-            dht_hash: holo_hash::AnyDhtHash,
-            options: holochain_p2p::event::GetMetaOptions,
-        ) -> BoxFut<'_, HolochainP2pResult<MetadataSet>> {
-            Box::pin(async { unimplemented!() })
-        }
-
-        fn handle_get_links(
-            &self,
-            dna_hash: DnaHash,
-            to_agent: AgentPubKey,
-            link_key: WireLinkKey,
-            options: holochain_p2p::event::GetLinksOptions,
-        ) -> BoxFut<'_, HolochainP2pResult<WireLinkOps>> {
-            Box::pin(async { unimplemented!() })
-        }
-
-        fn handle_count_links(
-            &self,
-            dna_hash: DnaHash,
-            to_agent: AgentPubKey,
-            query: WireLinkQuery,
-        ) -> BoxFut<'_, HolochainP2pResult<CountLinksResponse>> {
-            Box::pin(async { unimplemented!() })
-        }
-
-        fn handle_get_agent_activity(
-            &self,
-            dna_hash: DnaHash,
-            to_agent: AgentPubKey,
-            agent: AgentPubKey,
-            query: ChainQueryFilter,
-            options: holochain_p2p::event::GetActivityOptions,
-        ) -> BoxFut<'_, HolochainP2pResult<AgentActivityResponse>> {
-            Box::pin(async { unimplemented!() })
-        }
-
-        fn handle_must_get_agent_activity(
-            &self,
-            dna_hash: DnaHash,
-            to_agent: AgentPubKey,
-            author: AgentPubKey,
-            filter: holochain_zome_types::chain::ChainFilter,
-        ) -> BoxFut<'_, HolochainP2pResult<MustGetAgentActivityResponse>> {
-            Box::pin(async { unimplemented!() })
-        }
-
-        fn handle_validation_receipts_received(
-            &self,
-            dna_hash: DnaHash,
-            to_agent: AgentPubKey,
-            receipts: ValidationReceiptBundle,
-        ) -> BoxFut<'_, HolochainP2pResult<()>> {
-            Box::pin(async { unimplemented!() })
-        }
-
-        fn handle_countersigning_session_negotiation(
-            &self,
-            dna_hash: DnaHash,
-            to_agent: AgentPubKey,
-            message: holochain_p2p::event::CountersigningSessionNegotiationMessage,
-        ) -> BoxFut<'_, HolochainP2pResult<()>> {
-            Box::pin(async { unimplemented!() })
         }
     }
 }
