+++ crates/holochain/src/conductor/conductor/hc_p2p_handler_impl.rs
@@ -1,0 +1,103 @@
+use super::*;
+use holochain_p2p::HolochainP2pResult;
+use kitsune2_api::BoxFut;
+
+impl holochain_p2p::event::HcP2pHandler for Conductor {
+    fn handle_call_remote(
+        &self,
+        dna_hash: DnaHash,
+        to_agent: AgentPubKey,
+        zome_call_params_serialized: ExternIO,
+        signature: Signature,
+    ) -> BoxFut<'_, HolochainP2pResult<SerializedBytes>> {
+        Box::pin(async { unimplemented!() })
+    }
+
+    fn handle_publish(
+        &self,
+        dna_hash: DnaHash,
+        request_validation_receipt: bool,
+        countersigning_session: bool,
+        ops: Vec<holochain_types::dht_op::DhtOp>,
+    ) -> BoxFut<'_, HolochainP2pResult<()>> {
+        Box::pin(async { unimplemented!() })
+    }
+
+    fn handle_get(
+        &self,
+        dna_hash: DnaHash,
+        to_agent: AgentPubKey,
+        dht_hash: holo_hash::AnyDhtHash,
+        options: holochain_p2p::event::GetOptions,
+    ) -> BoxFut<'_, HolochainP2pResult<WireOps>> {
+        Box::pin(async { unimplemented!() })
+    }
+
+    fn handle_get_meta(
+        &self,
+        dna_hash: DnaHash,
+        to_agent: AgentPubKey,
+        dht_hash: holo_hash::AnyDhtHash,
+        options: holochain_p2p::event::GetMetaOptions,
+    ) -> BoxFut<'_, HolochainP2pResult<MetadataSet>> {
+        Box::pin(async { unimplemented!() })
+    }
+
+    fn handle_get_links(
+        &self,
+        dna_hash: DnaHash,
+        to_agent: AgentPubKey,
+        link_key: WireLinkKey,
+        options: holochain_p2p::event::GetLinksOptions,
+    ) -> BoxFut<'_, HolochainP2pResult<WireLinkOps>> {
+        Box::pin(async { unimplemented!() })
+    }
+
+    fn handle_count_links(
+        &self,
+        dna_hash: DnaHash,
+        to_agent: AgentPubKey,
+        query: WireLinkQuery,
+    ) -> BoxFut<'_, HolochainP2pResult<CountLinksResponse>> {
+        Box::pin(async { unimplemented!() })
+    }
+
+    fn handle_get_agent_activity(
+        &self,
+        dna_hash: DnaHash,
+        to_agent: AgentPubKey,
+        agent: AgentPubKey,
+        query: ChainQueryFilter,
+        options: holochain_p2p::event::GetActivityOptions,
+    ) -> BoxFut<'_, HolochainP2pResult<AgentActivityResponse>> {
+        Box::pin(async { unimplemented!() })
+    }
+
+    fn handle_must_get_agent_activity(
+        &self,
+        dna_hash: DnaHash,
+        to_agent: AgentPubKey,
+        author: AgentPubKey,
+        filter: holochain_zome_types::chain::ChainFilter,
+    ) -> BoxFut<'_, HolochainP2pResult<MustGetAgentActivityResponse>> {
+        Box::pin(async { unimplemented!() })
+    }
+
+    fn handle_validation_receipts_received(
+        &self,
+        dna_hash: DnaHash,
+        to_agent: AgentPubKey,
+        receipts: ValidationReceiptBundle,
+    ) -> BoxFut<'_, HolochainP2pResult<()>> {
+        Box::pin(async { unimplemented!() })
+    }
+
+    fn handle_countersigning_session_negotiation(
+        &self,
+        dna_hash: DnaHash,
+        to_agent: AgentPubKey,
+        message: holochain_p2p::event::CountersigningSessionNegotiationMessage,
+    ) -> BoxFut<'_, HolochainP2pResult<()>> {
+        Box::pin(async { unimplemented!() })
+    }
+}
